Eatery Hours Setup
Just need to fill in the different information for all of the eateries.

+++ STEP App/SWRevealViewController.m
@@ -1349,7 +1349,7 @@ const int FrontViewPositionNone = 0xff;
     
     NSTimeInterval duration = animated?_toggleAnimationDuration:0.0;
     NSTimeInterval firstDuration = duration;
-    int initialPosDif = labs( _frontViewPosition - preReplacementPosition );
+    int initialPosDif = (int) labs( _frontViewPosition - preReplacementPosition );
     if ( initialPosDif == 1 ) firstDuration *= 0.8;
     else if ( initialPosDif == 0 ) firstDuration = 0;
